requirements.txt
tensorboard
tqdm
torch-tb-profiler
matplotlib
numpy